requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pydantic==2.10.6
python-dotenv==1.0.1
google-genai==1.4.0
google-adk==0.1.0
httpx==0.28.1
httpx==0.28.1
pytest==8.3.5
